requirements.txt
sym-api-client-python>=1.1.3
requests
datetime
python-jose